alloy_sol_type_parser

Function is_valid_identifier

pub const fn is_valid_identifier(s: &str) -> bool

Returns true if the given string is a valid Solidity identifier.

An identifier in Solidity has to start with a letter, a dollar-sign or an underscore and may additionally contain numbers after the first symbol.

Solidity reference: https://docs.soliditylang.org/en/latest/grammar.html#a4.SolidityLexer.Identifier